+++ src/main.h
@@ -19,7 +19,7 @@
 #pragma once
 
 #ifdef BUILD_REL
-    #define MOD_VERSION "Thinker Mod v4.6 - The Will to Power mod v361"
+    #define MOD_VERSION "Thinker Mod v4.6 - The Will to Power mod v362"
 #else
     #define MOD_VERSION "Thinker Mod develop build"
 #endif